Refactor: rename package metric to quantity
