+++ src/main/java/de/fakeller/performance/analysis/result/unit/ThroughputUnit.java
@@ -1,6 +1,6 @@
 package de.fakeller.performance.analysis.result.unit;
 
-import de.fakeller.performance.analysis.result.metric.RelativeThroughput;
+import de.fakeller.performance.analysis.result.quantity.RelativeThroughput;
 
 public enum ThroughputUnit implements Unit<RelativeThroughput> {
     MESSAGES,
