+++ src/main/java/de/fakeller/performance/analysis/result/unit/TimeUnit.java
@@ -1,6 +1,6 @@
 package de.fakeller.performance.analysis.result.unit;
 
-import de.fakeller.performance.analysis.result.metric.TimeQuantity;
+import de.fakeller.performance.analysis.result.quantity.TimeQuantity;
 
 /**
  * Time units for {@link TimeQuantity} performance quantities.
