+++ src/main/java/de/fakeller/performance/analysis/result/unit/Unit.java
@@ -1,6 +1,6 @@
 package de.fakeller.performance.analysis.result.unit;
 
-import de.fakeller.performance.analysis.result.metric.PerformanceQuantity;
+import de.fakeller.performance.analysis.result.quantity.PerformanceQuantity;
 
 public interface Unit<Q extends PerformanceQuantity<?>> {
 }
